+++ spec/Suite/psr7/UriSpec.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 use Kahlan\Plugin\Stub;
 use Psr\Http\Message\UriInterface;